Fix latent PKCS12 dual-alias fragility; move cert helpers to shared package
The emulator's PFX carried a redundant cert-only alias (clientCert)
alongside the key alias (client) that WireMock/Kestrel actually need.
Empirically verified (a) the SunJSSE PKCS12KeyStore enumerates key
entries before cert-only entries regardless of alias name or insertion
order on this JDK, so the reused dsl/tls/support/CertificateHelpers'
positional certificates().get(0)/keys().get(0) pairing wasn't actually
broken in practice, but (b) it was relying on that undocumented
provider-internal ordering rather than an API guarantee, and this PR
was the first caller to feed that helper a multi-alias PKCS12. Drop the
redundant setCertificateEntry call - TrustManagerFactory already
derives a trust anchor from a key entry's own chain (verified), so
nothing depended on the separate cert-only alias.

Also move CertificateHelpers/SelfSignedCertificate out of the
hashicorp-specific dsl.certificates package into
tech.pegasys.web3signer.keystore.dsl.certificates, since they're now
shared fixtures used by both Hashicorp and Azure emulator acceptance
tests, not something hashicorp-specific.
