+++ keystorage/src/testFixtures/java/tech/pegasys/web3signer/keystore/dsl/certificates/CertificateHelpers.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.web3signer.keystore.hashicorp.dsl.certificates;
+package tech.pegasys.web3signer.keystore.dsl.certificates;
 
 import java.io.FileOutputStream;
 import java.io.IOException;
@@ -78,7 +78,6 @@ public class CertificateHelpers {
     ks.load(null, null);
 
     final Certificate certificate = selfSignedCert.getCertificate();
-    ks.setCertificateEntry("clientCert", certificate);
     ks.setKeyEntry(
         "client",
         selfSignedCert.getKeyPair().getPrivate(),
